add service implementation

+++ src/main/java/com/example/BalanceSheet/repository/ExpenseRepository.java
@@ -3,5 +3,5 @@ package com.example.BalanceSheet.repository;
 import com.example.BalanceSheet.model.Expense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpenseRepository extends JpaRepository<Expense, Long> {
+public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
 }
