+++ src/main/java/com/example/BalanceSheet/repository/IncomeRepository.java
@@ -3,5 +3,5 @@ package com.example.BalanceSheet.repository;
 import com.example.BalanceSheet.model.Income;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IncomeRepository extends JpaRepository<Income, Long> {
+public interface IncomeRepository extends JpaRepository<Income, Integer> {
 }
